CPLD: Extract library code from CRC tool.
